dom/canvas/test/test_filter_opacity_a8.html

<!DOCTYPE HTML>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body onload="runTest()">
<svg style="display: block; width: 0; height: 0">
  <defs>
    <filter id="black-flood">
      <feFlood flood-color="black" flood-opacity="1"/>
    </filter>
  </defs>
</svg>
<canvas id="c" width="10" height="4"></canvas>
<script>
function runTest() {
  var canvas = document.getElementById("c");
  var ctx = canvas.getContext("2d");
  const W = canvas.width;
  const H = canvas.height;

  ctx.filter = "url(#black-flood) opacity(50%)";
  ctx.fillRect(0, 0, W, H);

  var data = ctx.getImageData(0, 0, W, H).data;
  const expectedAlpha = 127;

  for (var y = 0; y < H; y++) {
    for (var x = 0; x < W; x++) {
      var alpha = data[(y * W + x) * 4 + 3];
      is(alpha, expectedAlpha, `pixel (${x},${y}) alpha should be ${expectedAlpha}`);
    }
  }

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
